--- app/src/main/java/com/example/petadoptionapp/PetIdClass.kt
@@ -1,3 +1,0 @@
-package com.example.petadoptionapp
-
-data class PetIdClass(var petId:Int)
