+++ app/src/main/java/com/example/petadoptionapp/activities/DeleteAccountActivity.kt
@@ -1,4 +1,4 @@
-package com.example.petadoptionapp
+package com.example.petadoptionapp.activities
 
 import android.app.Dialog
 import android.content.Context
@@ -7,6 +7,8 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.Window
 import android.widget.TextView
+import com.example.petadoptionapp.R
+import com.example.petadoptionapp.httpservice.MyPetAppApi
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
