+++ app/src/main/java/com/example/petadoptionapp/activities/PetInfoActivity.kt
@@ -1,16 +1,17 @@
-package com.example.petadoptionapp
+package com.example.petadoptionapp.activities
 
 import android.app.Dialog
-import android.content.Intent
-import android.content.SharedPreferences
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import android.view.Window
-import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
 import android.widget.Toast
+import com.example.petadoptionapp.PetsModel
+import com.example.petadoptionapp.R
+import com.example.petadoptionapp.data.PetIdClass
+import com.example.petadoptionapp.httpservice.MyPetAppApi
 import com.squareup.picasso.Picasso
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
