+++ app/src/main/java/com/example/petadoptionapp/data/InterestedInClass.kt
