+++ app/src/main/java/com/example/petadoptionapp/data/IpResult.kt
@@ -1,4 +1,4 @@
-package com.example.petadoptionapp
+package com.example.petadoptionapp.data
 
 data class IpResult(
     var id: Int = 0,
