+++ app/src/main/java/com/example/petadoptionapp/data/PetIdClass.kt
@@ -1,0 +1,3 @@
+package com.example.petadoptionapp.data
+
+data class PetIdClass(var petId:Int)
