+++ app/src/main/java/com/example/petadoptionapp/data/User.kt
@@ -1,3 +1,3 @@
-package com.example.petadoptionapp
+package com.example.petadoptionapp.data
 
 data class User(var email:String, var password:String)
