+++ app/src/main/res/layout/activity_all_pets_list.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:openDrawer="start"
-    tools:context=".AllPetsList">
+    tools:context=".activities.AllPetsList">
 
 
     <androidx.recyclerview.widget.RecyclerView
