+++ app/src/main/res/layout/activity_change_email.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ChangeEmailActivity">
+    tools:context=".activities.ChangeEmailActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/changeEmailInput"
