+++ app/src/main/res/layout/activity_delete_account.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DeleteAccountActivity">
+    tools:context=".activities.DeleteAccountActivity">
 
     <TextView
         android:id="@+id/accountDetails"
