+++ app/src/main/res/layout/activity_other_users.xml
@@ -7,7 +7,7 @@
     android:layout_marginStart="8dp"
     android:layout_marginTop="16dp"
     android:layout_marginEnd="8dp"
-    tools:context=".OtherUsersActivity">
+    tools:context=".activities.OtherUsersActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/otherUsers"
