+++ app/src/main/res/layout/activity_pet_info.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".PetInfoActivity">
+        tools:context=".activities.PetInfoActivity">
 
         <TextView
             android:id="@+id/petInfoHeadView"
