+++ app/src/main/res/layout/activity_start.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StartActivity">
+    tools:context=".activities.StartActivity">
 
     <ImageView
         android:id="@+id/startImageView"
